Refactor label utility functions
Moves some utility functions from `dask_image.ndmeasure`'s `__init__`
used to perform `label` over multiple chunks to `_utils._label`.
